Opting out of `-o default -o bashdefault` in generating completion script (#3)
* Opting out of `-o default -o bashdefault` in generating completion script

This is currently work in progress. It works, but it is rather cumbersome, since one needs to manually add a flag to the source statement. A better appraoch would be to have an option when registering the completion, but I cannot seem to move values from the Kong parser object to the final Kong command.

* Slightly better approach using two different commands

* Use annotation for configuration

* Remove obsolete struct

---------

Co-authored-by: Jan Heuermann <jan@jotaen.net>